More work on automating releases

+++ Build/python/mtm/zen/CreateRelease.py
@@ -13,6 +13,7 @@ from mtm.util.SystemHelper import SystemHelper
 from mtm.util.VarManager import VarManager
 from mtm.config.Config import Config
 from mtm.log.Logger import Logger
+from mtm.util.VisualStudioHelper import VisualStudioHelper
 
 from mtm.util.Assert import *
 
@@ -21,20 +22,24 @@ from mtm.ioc.Inject import Inject
 
 ScriptDir = os.path.dirname(os.path.realpath(__file__))
 RootDir = os.path.realpath(os.path.join(ScriptDir, '../../../..'))
+BuildDir = os.path.realpath(os.path.join(RootDir, 'Build'))
 
 class Runner:
     _log = Inject('Logger')
     _sys = Inject('SystemHelper')
     _varMgr = Inject('VarManager')
     _zipHelper = Inject('ZipHelper')
+    _vsSolutionHelper = Inject('VisualStudioHelper')
 
     def run(self):
         self._varMgr.add('RootDir', RootDir)
-        self._varMgr.add('BuildDir', '[RootDir]/Build')
+        self._varMgr.add('BuildDir', BuildDir)
         self._varMgr.add('PythonDir', '[BuildDir]/python')
         self._varMgr.add('TempDir', '[BuildDir]/Temp')
         self._varMgr.add('AssetsDir', '[RootDir]/UnityProject/Assets')
         self._varMgr.add('ZenjectDir', '[AssetsDir]/Zenject')
+        self._varMgr.add('DistDir', '[BuildDir]/Dist')
+        self._varMgr.add('BinDir', '[RootDir]/AssemblyBuild/Bin')
 
         versionStr = self._sys.readFileAsText('[BuildDir]/Version.txt').strip()
 
@@ -47,51 +52,95 @@ class Runner:
 
         minorNumber += 1
 
-        self._createReleaseZip(versionStr)
+        self._populateDistDir(versionStr)
 
-        self._sys.executeAndReturnOutput("git tag -a v{0}.{1} -m 'Version {0}.{1}'".format(majorNumber, minorNumber))
-        self._sys.writeFileAsText('[BuildDir]/Version.txt', '{0}.{1}'.format(majorNumber, minorNumber))
+        #self._sys.executeAndReturnOutput("git tag -a v{0}.{1} -m 'Version {0}.{1}'".format(majorNumber, minorNumber))
+        #self._sys.writeFileAsText('[BuildDir]/Version.txt', '{0}.{1}'.format(majorNumber, minorNumber))
 
         self._log.info("Incremented version to {0}.{1}. Now commit and then run 'git push --tags'\n\n", majorNumber, minorNumber)
 
-    def _createReleaseZip(self, versionStr):
-        self._log.heading('Creating release zip file')
+    def _populateDistDir(self, versionStr):
+
+        self._sys.createDirectory('[DistDir]')
+        self._sys.clearDirectoryContents('[DistDir]')
+
+        self._createPackage(False, True, '[DistDir]/Zenject-{0}-WithAsteroidsDemo.unitypackage'.format(versionStr))
+        self._createPackage(False, False, '[DistDir]/Zenject-{0}.unitypackage'.format(versionStr))
+        self._createPackage(True, False, '[DistDir]/Zenject-{0}-BinariesOnly.unitypackage'.format(versionStr))
+
+        self._createNonUnityZip('[DistDir]/Zenject-{0}-NonUnity.zip'.format(versionStr))
+
+    def _createNonUnityZip(self, zipPath):
+
+        self._log.heading('Creating non unity zip')
+
+        tempDir = '[TempDir]/NonUnity'
+        self._sys.createDirectory(tempDir)
+        self._sys.clearDirectoryContents(tempDir)
+
+        binDir = '[BinDir]/Not Unity Release'
+        self._sys.clearDirectoryContents(binDir)
+        self._vsSolutionHelper.buildVisualStudioProject('[RootDir]/AssemblyBuild/Zenject.sln', 'Not Unity Release')
+
+        self._log.info('Copying Zenject dlls')
+        self._sys.copyFile('{0}/Zenject.dll'.format(binDir), '{0}/Zenject.dll'.format(tempDir))
+        self._sys.copyFile('{0}/Zenject.Commands.dll'.format(binDir), '{0}/Zenject.Commands.dll'.format(tempDir))
+
+        self._zipHelper.createZipFile(tempDir, zipPath)
+
+    def _createPackage(self, useDll, includeSample, outputPath):
+
+        self._log.heading('Creating {0}'.format(os.path.basename(outputPath)))
 
         self._sys.createDirectory('[TempDir]')
         self._sys.clearDirectoryContents('[TempDir]')
 
         self._varMgr.add('ZenTempDir', '[TempDir]/Packager/Assets/Zenject')
 
-        self._log.info('Copying Zenject to temporary directory')
-        self._sys.copyDirectory('[ZenjectDir]', '[ZenTempDir]')
+        if useDll:
+            self._log.info('Building zenject dlls')
+            self._sys.clearDirectoryContents('[BinDir]/Release')
+            self._vsSolutionHelper.buildVisualStudioProject('[RootDir]/AssemblyBuild/Zenject.sln', 'Release')
 
-        self._log.info('Cleaning up Zenject directory')
-        self._zipHelper.createZipFile('[ZenTempDir]/Extras/ZenjectUnitTests', '[ZenTempDir]/Extras/ZenjectUnitTests.zip')
-        self._sys.deleteDirectory('[ZenTempDir]/Extras/ZenjectUnitTests')
+            self._log.info('Copying Zenject dlls')
+            self._sys.copyFile('[BinDir]/Release/Zenject.dll', '[ZenTempDir]/Zenject.dll')
+            self._sys.copyFile('[BinDir]/Release/Zenject-editor.dll', '[ZenTempDir]/Editor/Zenject-editor.dll')
+            self._sys.copyFile('[BinDir]/Release/Zenject.Commands.dll', '[ZenTempDir]/Zenject.Commands.dll')
+        else:
+            self._log.info('Copying Zenject to temporary directory')
+            self._sys.copyDirectory('[ZenjectDir]', '[ZenTempDir]')
 
-        self._zipHelper.createZipFile('[ZenTempDir]/Extras/ZenjectAutoMocking', '[ZenTempDir]/Extras/ZenjectAutoMocking.zip')
-        self._sys.deleteDirectory('[ZenTempDir]/Extras/ZenjectAutoMocking')
+            self._log.info('Cleaning up Zenject directory')
+            self._zipHelper.createZipFile('[ZenTempDir]/Extras/ZenjectUnitTests', '[ZenTempDir]/Extras/ZenjectUnitTests.zip')
+            self._sys.deleteDirectory('[ZenTempDir]/Extras/ZenjectUnitTests')
+
+            self._zipHelper.createZipFile('[ZenTempDir]/Extras/ZenjectAutoMocking', '[ZenTempDir]/Extras/ZenjectAutoMocking.zip')
+            self._sys.deleteDirectory('[ZenTempDir]/Extras/ZenjectAutoMocking')
+
+            if not includeSample:
+                self._sys.deleteDirectory('[ZenTempDir]/Extras/SampleGame')
 
         self._sys.createDirectory('[TempDir]/Packager/ProjectSettings')
+        self._sys.createDirectory('[TempDir]/Packager/Assets')
 
         self._sys.copyFile('[BuildDir]/UnityPackager/UnityPackageUtil.cs', '[TempDir]/Packager/Assets/Editor/UnityPackageUtil.cs')
 
-        self._log.info('Creating unity package')
+        self._log.info('Running unity to create unity package')
         self._sys.executeAndWait('"[UnityExePath]" -batchmode -nographics -quit -projectPath "[TempDir]/Packager" -executeMethod Zenject.UnityPackageUtil.CreateUnityPackage')
 
-        self._varMgr.add('DistDir', '[BuildDir]/Dist')
-
-        self._sys.createDirectory('[DistDir]')
-        self._sys.clearDirectoryContents('[DistDir]')
-
-        self._sys.copyFile('[TempDir]/Packager/Zenject.unitypackage', '[DistDir]/Zenject-{0}.unitypackage'.format(versionStr))
+        self._sys.move('[TempDir]/Packager/Zenject.unitypackage', outputPath)
 
 def installBindings():
 
     config = {
         'PathVars': {
             'UnityExePath': 'C:/Program Files/Unity/Editor/Unity.exe',
-        }
+            'LogPath': os.path.join(BuildDir, 'Log.txt'),
+            'MsBuildExePath': 'C:/Windows/Microsoft.NET/Framework/v4.0.30319/msbuild.exe'
+        },
+        'Compilation': {
+            'UseDevenv': False
+        },
     }
     Container.bind('Config').toSingle(Config, [config])
 
@@ -103,6 +152,7 @@ def installBindings():
     Container.bind('Logger').toSingle(Logger)
     Container.bind('ProcessRunner').toSingle(ProcessRunner)
     Container.bind('ZipHelper').toSingle(ZipHelper)
+    Container.bind('VisualStudioHelper').toSingle(VisualStudioHelper)
 
 if __name__ == '__main__':
